+++ lib/cims/lib/request/get-incident-request.ts
@@ -23,25 +23,29 @@ export interface GetIncidentRequest extends common.BaseRequest {
    */
   "incidentKey": string;
   /**
+   * The OCID of the tenancy.
+   */
+  "compartmentId": string;
+  /**
    * Unique Oracle-assigned identifier for the request. If you need to contact Oracle about a particular request, please provide the request ID.
    */
   "opcRequestId"?: string;
   /**
-   * The Customer Support Identifier (CSI) associated with the support account.
+   * The Customer Support Identifier (CSI) number associated with the support account.
+   * The CSI is required for technical support tickets and optional for limits and billing tickets.
+   *
    */
   "csi"?: string;
   /**
    * User OCID for Oracle Identity Cloud Service (IDCS) users who also have a federated Oracle Cloud Infrastructure account.
+   * User OCID is mandatory for OCI Users and optional for Multicloud users.
+   *
    */
   "ocid"?: string;
   /**
    * The region of the tenancy.
    */
   "homeregion"?: string;
-  /**
-   * The OCID of the tenancy.
-   */
-  "compartmentId"?: string;
   /**
    * The kind of support request.
    */
@@ -60,6 +64,8 @@ export interface GetIncidentRequest extends common.BaseRequest {
   "idtoken"?: string;
   /**
    * The OCID of identity domain.
+   * DomainID is mandatory if the user is part of Non Default Identity domain.
+   *
    */
   "domainid"?: string;
 }