+++ lib/cims/lib/request/list-incidents-request.ts
@@ -23,7 +23,9 @@ export interface ListIncidentsRequest extends common.BaseRequest {
    */
   "compartmentId": string;
   /**
-   * The Customer Support Identifier (CSI) associated with the support account.
+   * The Customer Support Identifier (CSI) number associated with the support account.
+   * The CSI is required for technical support tickets and optional for limits and billing tickets.
+   *
    */
   "csi"?: string;
   /**
@@ -54,6 +56,8 @@ export interface ListIncidentsRequest extends common.BaseRequest {
   "opcRequestId"?: string;
   /**
    * User OCID for Oracle Identity Cloud Service (IDCS) users who also have a federated Oracle Cloud Infrastructure account.
+   * User OCID is mandatory for OCI Users and optional for Multicloud users.
+   *
    */
   "ocid"?: string;
   /**
@@ -78,6 +82,8 @@ export interface ListIncidentsRequest extends common.BaseRequest {
   "idtoken"?: string;
   /**
    * The OCID of identity domain.
+   * DomainID is mandatory if the user is part of Non Default Identity domain.
+   *
    */
   "domainid"?: string;
 }