+++ lib/cims/lib/request/validate-user-request.ts
@@ -19,7 +19,9 @@ import common = require("oci-common");
  */
 export interface ValidateUserRequest extends common.BaseRequest {
   /**
-   * The Customer Support Identifier (CSI) associated with the support account.
+   * The Customer Support Identifier (CSI) number associated with the support account.
+   * The CSI is required for technical support tickets and optional for limits and billing tickets.
+   *
    */
   "csi"?: string;
   /**
@@ -32,6 +34,8 @@ export interface ValidateUserRequest extends common.BaseRequest {
   "problemType"?: model.ProblemType;
   /**
    * User OCID for Oracle Identity Cloud Service (IDCS) users who also have a federated Oracle Cloud Infrastructure account.
+   * User OCID is mandatory for OCI Users and optional for Multicloud users.
+   *
    */
   "ocid"?: string;
   /**
@@ -52,6 +56,8 @@ export interface ValidateUserRequest extends common.BaseRequest {
   "idtoken"?: string;
   /**
    * The OCID of identity domain.
+   * DomainID is mandatory if the user is part of Non Default Identity domain.
+   *
    */
   "domainid"?: string;
 }